+++ Tasks/OrchestratorV3/extensions/buildapiretry/buildapiretry.ts
@@ -3,14 +3,18 @@ import { Build, BuildReason, BuildStatus, BuildResult, QueryDeletedOption, Build
 
 import { IBuildApiRetry } from "./ibuildapiretry";
 import { Retryable } from "../../common/retry";
+import { IApiClient } from "../../common/iapiclient";
+import { IRestResponse } from "typed-rest-client";
 
 export class BuildApiRetry implements IBuildApiRetry {
 
     private buildApi: IBuildApi;
+    private apiClient: IApiClient;
 
-    constructor(buildApi: IBuildApi) {
+    constructor(buildApi: IBuildApi, apiClient: IApiClient) {
 
         this.buildApi = buildApi;
+        this.apiClient = apiClient;
 
     }
 
@@ -125,14 +129,22 @@ export class BuildApiRetry implements IBuildApiRetry {
 
     }
 
+    // BuildApi uses old API version which does not support forceRetryAllJobs parameter
+    // Therefore using our own implementation to make updateStage REST call
     @Retryable()
     public async updateStage(updateParameters: UpdateStageParameters, buildId: number, stageRefName: string, project?: string): Promise<void> {
 
-        return await this.buildApi.updateStage(
-            updateParameters,
-            buildId,
-            stageRefName,
-            project);
+        const run: unknown = await this.apiClient.patch(`${project}/_apis/build/builds/${buildId}/stages/${stageRefName}`, `7.1-preview.1`, updateParameters, true);
+
+        const responseCode: number | undefined = (run as IRestResponse<number>).statusCode;
+
+        const validResponseCodes: number[] = [ 200, 204 ];
+
+        if (!responseCode || !validResponseCodes.includes(responseCode)) {
+
+            throw new Error(`Unable to update <${stageRefName}> stage status`);
+
+        }
 
     }